Reorganization (move experiment to characterization)
