--- docs/apidocs/mod_stark.rst
@@ -1,6 +1,0 @@
-.. _qiskit-experiments-stark:
-
-.. automodule:: qiskit_experiments.library.stark
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
