+++ qiskit_experiments/library/__init__.py
@@ -78,23 +78,7 @@ Some experiments may be also used for gate calibration.
     ~characterization.ResonatorSpectroscopy
     ~characterization.ZZRamsey
     ~characterization.MultiStateDiscrimination
-
-.. _stark:
-
-Stark Experiments
-=================
-
-Experiments with a Stark tone. When a qubit is driven with an off-resonant tone,
-the qubit frequency is slightly shifted through what is known as the Stark effect.
-This technique is sometime used to characterize qubit properties in the vicinity of
-the base frequency, especially with the fixed frequency qubit architecture which usually
-doesn't have a knob to control frequency.
-
-.. autosummary::
-    :toctree: ../stubs/
-    :template: autosummary/experiment.rst
-
-    ~stark.StarkRamseyXY
+    ~characterization.StarkRamseyXY
 
 .. _calibration:
 
@@ -168,11 +152,11 @@ from .characterization import (
     CorrelatedReadoutError,
     ZZRamsey,
     MultiStateDiscrimination,
+    StarkRamseyXY,
 )
 from .randomized_benchmarking import StandardRB, InterleavedRB
 from .tomography import StateTomography, ProcessTomography
 from .quantum_volume import QuantumVolume
-from .stark import StarkRamseyXY
 
 # Experiment Sub-modules
 from . import calibration
@@ -180,4 +164,3 @@ from . import characterization
 from . import randomized_benchmarking
 from . import tomography
 from . import quantum_volume
-from . import stark
